Crate libsignal_bridge_macros

Generates C, Java, and Node entry points for Rust functions.

The goal of the bridge_fn family of macros is to define a cross-language glue layer using strongly-typed Rust code. You can write a normal top-level Rust function exposing a particular operation, and additional functions will be generated for FFI, JNI, and Node bindings, mapping types automatically.

It is explicitly not a goal for this layer to generate perfect C, Java, or TypeScript APIs. Rather, it should generate safe interfaces to Rust APIs, on top of which idiomatic Swift, Java, and TypeScript APIs can be built.

Example

#[bridge_fn]
fn SenderKeyMessage_New(
    key_id: u32,
    iteration: u32,
    ciphertext: &[u8],
    pk: &PrivateKey,
) -> Result<SenderKeyMessage> {
    let mut csprng = rand::rngs::OsRng;
    SenderKeyMessage::new(key_id, iteration, ciphertext, &mut csprng, pk)
}

SignalFfiError *signal_sender_key_message_new(
    SignalSenderKeyMessage **out,
    uint32_t key_id,
    uint32_t iteration,
    const unsigned char *ciphertext,
    size_t ciphertext_len,
    const SignalPrivateKey *pk);

public static native long SenderKeyMessage_New(
    int keyId,
    int iteration,
    byte[] ciphertext,
    long pk);

export function SenderKeyMessage_New(
    keyId: number,
    iteration: number,
    ciphertext: Buffer,
    pk: Wrapper<PrivateKey>
): SenderKeyMessage;

Async support for Node

For the Node bridge, if a bridge_fn is declared as async, it will return a JavaScript Promise and run the function on the JavaScript event loop using the signal-neon-futures crate. Interaction with JavaScript can be done through async callbacks, including trait objects defined using the async-trait crate. Like the synchronous implementations of all three bridges, panics will be caught and translated to JavaScript exceptions.

The FFI and JNI bridges do not support asynchronous execution; an async function is invoked and expected to complete immediately without blocking.

Naming conventions

By default, bridge_fn tries to pick a good name for each exposed entry point:

FFI: Convert the function's name to lower_snake_case and prepend signal_.

JNI: Escape any underscores in the function's name per the JNI spec, then prepend Java_org_signal_client_internal_Native_ to expose the function as a static method of the class org.signal.client.internal.Native.

Node: Use the original function's name.

As such, the recommended naming scheme for bridge_fn functions is ObjectOrGroup_Operation.

Any of these names can be replaced by specifying an argument to the bridge_fn attribute:

#[bridge_fn(ffi = "magic_alakazam", jni = "Magic_1Alakazam")]
fn Abracadabra() {
  // ...
}

A replaced name does not undergo any transformation, but is still prefixed with the required "namespace" for FFI and JNI.

Limiting to certain bridges

Do not use cfg(feature = "abc") to restrict a bridge_fn to certain bridges (e.g. "just FFI"). This interacts poorly with commands like cargo clippy --workspace, which try to validate all packages by enabling all three bridges at once. Instead, you can write e.g. bridge_fn(jni = false) to keep from exposing a particular function to Java.

Adding new argument and result types

If your argument or result type is a Rust value being wrapped in an opaque box, declare it using the bridge_handle macro alongside other such types. Otherwise, there are two steps:

Argument and result types for FFI and JNI are determined by macros ffi_arg_type, ffi_result_type, jni_arg_type, and jni_result_type. You may need to add your new type there. JNI and Node types also undergo some additional transformation in the scripts gen_java_decl.py and gen_ts_decl.py, which you may need to tweak as well.

Argument types conform to one or more of the following bridge-specific traits:

ffi::ArgTypeInfo or ffi::SizedArgTypeInfo

jni::ArgTypeInfo

node::ArgTypeInfo and/or node::AsyncArgTypeInfo

Similarly, result types conform to one or more of the following:

ffi::ResultTypeInfo

jni::ResultTypeInfo

node::ResultTypeInfo

These traits define how to convert between the bridge type and the Rust type used in the function as written. See each individual trait for more info on how to add a new type.

Limitations

Input buffers require special treatment for FFI so that their size can be passed in. This needs special handling in the implementation of the macros to generate multiple parameters in the FFI entry point that map to a single parameter in the corresponding Rust function. Supporting more types that would require multiple parameters is non-trivial, particularly when trying to do so on the syntactic representation of the AST that macros are restricted to.

There is no support for multiple return values, even though some of the FFI entry points use multiple output parameters. These functions must be implemented manually.

Modules

ffi

jni

node

Enums

ResultKind

Functions

bridge_fn_impl

name_for_meta_key

value_for_meta_key

Attribute Macros

bridge_fn
Generates C, Java, and Node entry points for a Rust function that returns a value.

bridge_fn_buffer
Generates C, Java, and Node entry points for a Rust function that returns a buffer, avoiding unnecessary copies of the return value.

bridge_fn_void
Generates C, Java, and Node entry points for a Rust function that returns Result<(), _>.