Attribute Macro libsignal_bridge_macros::bridge_fn_void

#[bridge_fn_void]

Generates C, Java, and Node entry points for a Rust function that returns Result<(), _>.

Because the C bindings conventions use out-parameters for successful return values, a case of "no result on success" must be annotated specially.

See the crate-level documentation for more information.

Example

// Produces a C function manually named "signal_process_postkey"
// and a JNI function named "PostKey_1Process" (with JNI "_1" mangling for an underscore),
// with the Node entry point disabled.
#[bridge_fn_void(ffi = "process_postkey", node = false)]
fn PostKey_Process(post_key: &PostKey) -> Result<()> {
  // ...
}